Added unit tests and a formatted string framework.
Unit tests allow a hierarchy of test suites, each suite can have tests
and assertions. Tests log their status with either a pass or fail.
Assertions do nothing on success, and throw an exception on failure.
Exceptions are caught by RunSuite or RunTests and add a failure to the
the test that threw them.
The formatted string works similar to string stream, by concatenating
the string representation of up to 5 parameters.